% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_community.R
\name{diamond_community}
\alias{diamond_community}
\title{To calculate the gene abundance in microbial community (GAM,\%).}
\usage{
diamond_community(
  input_reads,
  result,
  threads = 1,
  SCG_db,
  USCG_db = "default",
  skip_fastp = F,
  min_length = 100,
  run_seqkit = T,
  filter_condition = "default",
  keep_samples = F
)
}
\arguments{
\item{input_reads}{Please set the reads to include only the forward reads if the data is paired-end (PE).}

\item{result}{Please set the name for the result file.}

\item{threads}{Set the number of CPU threads, with the default value being 1.}

\item{SCG_db}{Please specify the directory for the single-copy genes DIAMOND database.}

\item{USCG_db}{Please specify the directory for universal single-copy ribosomal genes database (e.g., 'Ribo_14.dmnd'). If you have already calculated the RPKM for these genes, you may instead specify the directory for the results (e.g., 'sample_name.USCG.hits.txt') to skip this step.}

\item{skip_fastp}{If you have already filtered the reads, you can set T to skip running fastp. The default is to run fastp.}

\item{min_length}{Set the minimum length required for filtering reads, the default is 100 bp.}

\item{run_seqkit}{If you have already counted the total number of reads using seqkit, you can specify the directory of the seqkit results (e.g., 'sample_name.all.reads.txt') to skip running seqkit. By default, seqkit will be executed.}

\item{filter_condition}{Please specify the filter_condition file. By default, the identity is set to 50 and the coverage to 80.}

\item{keep_samples}{By default, the temporary results will be deleted unless this parameter is setted.}
}
\value{
The gene abundance in microbial community (GAM,\%)
}
\description{
To calculate the gene abundance in microbial community (GAM,\%).
}
\examples{
diamond_community(input_reads = "your/reads/data/sample_1.fastq",result = "sample_1",threads = 40,diamond_db = "your/database/ter_hydB-aprA.dmnd",USCG_db = "your/database/Ribo_14.dmnd",min_length = 70,filter_condition = "default")
}
