% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd.diamond_community.R
\name{hyd.diamond_community}
\alias{hyd.diamond_community}
\title{To calculate hydrogenase's GAM.}
\usage{
hyd.diamond_community(
  input_reads,
  result,
  threads = 1,
  hyd_db = "default",
  script = "default",
  USCG_db = "default",
  skip_fastp = F,
  min_length = 100,
  run_seqkit = T,
  keep_samples = F
)
}
\arguments{
\item{input_reads}{Please set the reads to include only the forward reads if the data is paired-end (PE).}

\item{result}{Please specify the name for the result file.}

\item{threads}{Set the number of CPU threads, with the default value being 1.}

\item{hyd_db}{Please set the directory of single-copy genes' diamond database (e.g.'hyddb.all.dmnd').}

\item{script}{Please set the directory of hyd_id-name.script.}

\item{USCG_db}{Please specify the directory for universal single-copy genes database (e.g., 'Ribo_14.dmnd'). If you have already calculated the RPKM for these genes, you may instead specify the directory for the results (e.g., 'sample_name.USCG.hits.txt') to skip this step.}

\item{skip_fastp}{If you have already filtered the reads, you can set this parameter to skip running fastp. The default is to run fastp.}

\item{min_length}{Set the minimum length required for filtering reads, the default is 100.}

\item{run_seqkit}{If you have already counted the total number of reads using seqkit, you can specify the directory of the seqkit results (e.g., 'sample_name.all.reads.txt') to skip running seqkit. By default, seqkit will be executed.}

\item{keep_samples}{By default, the temporary results will be deleted unless this parameter is setted.}
}
\value{
The gene abundance in microbial community (GAM,\%)
}
\description{
To calculate hydrogenase's GAM.
}
\examples{
diamond_community(input_reads = "your/reads/data/sample_1.fastq",result = "sample_1",threads = 40,diamond_db = "your/database/hyddb.all.dmnd",USCG_db = "your/database/Ribo_14.dmnd",script = "your/database/hyd_id-name.script",min_length = 70)
}
