% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribo_rpkm.R
\name{uscg_rpkm}
\alias{uscg_rpkm}
\title{To calculate RPKM of universal single copy genes (RUSCG).}
\usage{
uscg_rpkm(
  input_reads,
  result,
  threads = 1,
  USCG_db,
  skip_fastp = F,
  min_length = 100,
  run_seqkit = T,
  keep_samples = F
)
}
\arguments{
\item{input_reads}{Please set the directory of reads}

\item{result}{Please set the directory of reads.}

\item{threads}{Setting the threads of CPU,default is 1.}

\item{USCG_db}{Please set the directory of universal single copy genes (USCGs) database.}

\item{skip_fastp}{If you have already filtered the reads, you can set this parameter to skip running fastp. The default is to run fastp.}

\item{min_length}{Set the minimum length required for filtering reads, the default is 100.}

\item{run_seqkit}{If you have already counted the total number of reads using seqkit, you can specify the directory of the seqkit results (e.g., 'sample_name.all.reads.txt') to skip running seqkit. By default, seqkit will be executed.}

\item{keep_samples}{By default, the temporary results will be deleted unless this parameter is setted.}
}
\value{
The rpkm abandance of universal single copy genes (RUSCG).
}
\description{
To calculate RPKM of universal single copy genes (RUSCG).
}
\examples{
ribo_rpkm(input_reads = "your/reads/data/sample_1.fastq",result = "sample_1",threads = 40,USCG_db = "your/database/Ribo_14.dmnd",skip_fastp = F,min_length = 70,run_seqkit = T,keep_samples = F)
}
